fix one integration test
