+++ package.json
@@ -26,7 +26,7 @@
     "es5-shim": "^4.0.3",
     "object-assign": "^2.0.0",
     "react": "~0.13.3",
-    "ron-react-autocomplete": "2.0.0"
+    "ron-react-autocomplete": "2.0.1"
   },
   "devDependencies": {
     "avcoveralls": "^1.0.0",
